backend_py/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
sentence-transformers==3.4.1
numpy==2.2.3
requests==2.32.3
python-dotenv==1.0.1
